old-tests/submission/Microsoft/async/async_009.htm

<!DOCTYPE html>
<html>
    <head>
        <title>Document.write() silently fails from an Async script</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta description="This test ensures Document.write() silently fails from an Async script." />
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
        <link rel="help" href="http://dev.w3.org/html5/spec/scripting-1.html#execute-the-script-block"/>

        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script type="text/javascript">           
            
        
  
        var t = async_test("Document.write() silently fails from an Async script");
  
        function timeout()
        {
            t.step(function(){ assert_true(!document.getElementById("writtenText"))});
            t.done();
        }
  
        var timer = setTimeout(timeout, 8000);
            
        function log(text)
        {
            document.write("<span id='failed'/>");
        }
        </script>
    </head>
    <body>
        <div id=log></div>
        <script type="text/javascript" src="log.py?sec=2&id=1" async></script>
    </body>
</html>
